Bug 1828008 - Avoid creating timers with a delay of UINT32_MAX in nsCaret::ResetBlinking, r=emilio.
Differential Revision: https://phabricator.services.mozilla.com/D175421

UltraBlame original commit: 05fc7d57257e06a4fe340966226f6764dfc3bcea

+++ layout/base/nsCaret.cpp
@@ -604,7 +604,7 @@ void nsCaret::ResetBlinking() {
   }
 
   auto blinkRate =
-      uint32_t(LookAndFeel::GetInt(IntID::CaretBlinkTime, kDefaultBlinkRate));
+      LookAndFeel::GetInt(IntID::CaretBlinkTime, kDefaultBlinkRate);
 
   if (blinkRate > 0) {
